+++ src/siena07models.cpp
@@ -333,7 +333,8 @@ SEXP forwardModel(SEXP DERIV, SEXP DATAPTR, SEXP SEEDS,
 						{
 							PutRNGstate();
 							SET_VECTOR_ELT(VECTOR_ELT(seedstore, group),
-								period, Rf_findVar(rs, R_GlobalEnv));
+//								period, Rf_findVar(rs, R_GlobalEnv)); misschien Rf_eval(rs, R_GlobalEnv)
+								period, R_getVar(rs, R_GlobalEnv, TRUE));
 						}
 					}
 				}
@@ -446,7 +447,8 @@ SEXP forwardModel(SEXP DERIV, SEXP DATAPTR, SEXP SEEDS,
 
 	/* send the .Random.seed back to R */
 	PutRNGstate();
-	NEWRANDOMSEED = Rf_findVar(rs, R_GlobalEnv);
+//	NEWRANDOMSEED = Rf_findVar(rs, R_GlobalEnv); misschien Rf_eval(rs, R_GlobalEnv)
+NEWRANDOMSEED = R_getVar(rs, R_GlobalEnv, TRUE);
 
 	/* set up the return object */
 	if (!fromFiniteDiff)